requirements.txt
Django==5.2.5
djangorestframework==3.15.2
django-cors-headers==4.4.0
Pillow==10.4.0
gunicorn==23.0.0
whitenoise==6.5.0
Flask==3.0.0


django-filter


